Update sam1.md
Aligned the text. 
I've encountered this entry multiple times today, and the text iscontinuous to the right of the screen which you need to scroll, hopefully the original creator will understand the changes its just to show properly. If you need to confirm 1st i've taken a screenshot of this specific entry ingame.